+++ pyppeteer_ghost_cursor/playwright/async_api/__init__.py
@@ -1,4 +1,4 @@
-from .spoof import create_cursor
-from .mouse_helper import install_mouse_helper
+from ._spoof import create_cursor
+from ._mouse_helper import install_mouse_helper
 
 __all__ = ["create_cursor", "install_mouse_helper"]
