+++ pyppeteer_ghost_cursor/playwright/async_api/_mouse_helper.py
