+++ pyppeteer_ghost_cursor/playwright/sync_api/_mouse_helper.py
