+++ pyppeteer_ghost_cursor/shared/__init__.py
@@ -1,3 +1,3 @@
-from .spoof import get_path as path
+from ._spoof import get_path as path
 
 __all__ = ["path"]
