+++ pyppeteer_ghost_cursor/shared/_spoof.py
@@ -2,7 +2,7 @@ import math
 import random
 import numpy as np
 from typing import Union, Optional, Dict, List
-from pyppeteer_ghost_cursor.shared.math import (
+from pyppeteer_ghost_cursor.shared._math import (
     Vector,
     magnitude,
     direction,
